rename plugins/bundleSize.js, remove shelljs dep
